requirements.txt
flask>=3.0.0
TTS>=0.22.0
flask_cors
google-generativeai
python-dotenv